move file to package xml
